refactor(bin): introduce server/http3.rs and server/http09.rs
The QUIC Interop Runner requires an http3 and http09 implementation for both
client and server. The client code is already structured into an http3 and an
http09 implementation since https://github.com/mozilla/neqo/pull/1727.

This commit does the same for the server side, i.e. splits the http3 and http09
implementation into separate Rust modules.
